cant print weights with these parameters

+++ include/Model.hpp
@@ -15,7 +15,7 @@ class Model {
     std::vector<std::vector<float>*> nablaCache; // holds gradient for each activated value
     std::vector<std::vector<std::vector<float>*>*> nablaWeights; // holds gradient for each weight
     std::vector<std::vector<std::vector<float>*>*> weights;
-    long double summationLoss;
+    float summationLoss;
     float learningRate;
 
     // initialize model members
